moved client integration tests to tests folder
